requirements.txt
# Core runtime
python-dotenv==1.0.1
pydantic==2.6.1

# Backend
fastapi==0.110.0
uvicorn[standard]==0.29.0

# Frontend
streamlit==1.32.0

# Data & Finance
pandas==2.2.1
numpy==1.26.4
yfinance==0.2.37
nsepy==0.8

# HTTP + Utilities
requests==2.31.0

# Logging helpers
loguru==0.7.2

-e .